pandam-first-client-lambda-func/requirements.txt
aws-cdk-lib==2.175.0
constructs>=10.0.0,<11.0.0
python-dotenv
watchdog